requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
amqp==2.5.0
argon2-cffi==19.1.0
async-timeout==3.0.1
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-compat==1.0.15
django-cors-headers==2.4.0
django-decouple==2.1
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-rest-coreapi-schema==0.1.2
django-timezone-field==3.0
djangorestframework==3.7.7
dnspython==1.16.0
eventlet==0.25.0
greenlet==0.4.15
idna==2.8
idna-ssl==1.1.0
itypes==1.1.0
Jinja2==2.10
kombu==4.6.3
Markdown==2.6.10
MarkupSafe==1.1.1
monotonic==1.5
multidict==4.5.2
oauthlib==3.0.1
pycares==3.0.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
python-memcached==1.59
pytz==2018.9
PyYAML==3.13
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
typing==3.7.4
typing-extensions==3.7.4
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.24.1
user-agents==1.1.0
vine==1.3.0
yarl==1.3.0
